+++ tests/Unit/RuleSet/DeclarationBlockTest.php
@@ -8,7 +8,11 @@ use PHPUnit\Framework\TestCase;
 use Sabberworm\CSS\CSSElement;
 use Sabberworm\CSS\CSSList\CSSListItem;
 use Sabberworm\CSS\Position\Positionable;
+use Sabberworm\CSS\Parsing\ParserState;
+use Sabberworm\CSS\Property\Selector;
 use Sabberworm\CSS\RuleSet\DeclarationBlock;
+use Sabberworm\CSS\Settings;
+use TRegx\PhpUnit\DataProviders\DataProvider;
 
 /**
  * @covers \Sabberworm\CSS\RuleSet\DeclarationBlock
@@ -47,5 +51,69 @@ final class DeclarationBlockTest extends TestCase
     public function implementsPositionable(): void
     {
         self::assertInstanceOf(Positionable::class, $this->subject);
+    }
+
+    /**
+     * @return array<non-empty-string, array{0: non-empty-string}>
+     */
+    public static function provideSelector(): array
+    {
+        return [
+            'type' => ['body'],
+            'class' => ['.teapot'],
+            'id' => ['#my-mug'],
+            '`not`' => [':not(#your-mug)'],
+            '`not` with multiple arguments' => [':not(#your-mug, .their-mug)'],
+        ];
+    }
+
+    /**
+     * @test
+     *
+     * @param non-empty-string $selector
+     *
+     * @dataProvider provideSelector
+     */
+    public function parsesAndReturnsSingleSelector(string $selector): void
+    {
+        $subject = DeclarationBlock::parse(new ParserState($selector . '{}', Settings::create()));
+
+        $resultSelectorStrings = \array_map(
+            static function (Selector $selectorObject): string {
+                return $selectorObject->getSelector();
+            },
+            $subject->getSelectors()
+        );
+        self::assertSame([$selector], $resultSelectorStrings);
+    }
+
+    /**
+     * @return DataProvider<non-empty-string, array{0: non-empty-string, 1: non-empty-string}>
+     */
+    public static function provideTwoSelectors(): DataProvider
+    {
+        return DataProvider::cross(self::provideSelector(), self::provideSelector());
+    }
+
+    /**
+     * @test
+     *
+     * @param non-empty-string $firstSelector
+     * @param non-empty-string $secondSelector
+     *
+     * @dataProvider provideTwoSelectors
+     */
+    public function parsesAndReturnsTwoCommaSeparatedSelectors(string $firstSelector, string $secondSelector): void
+    {
+        $joinedSelectors = $firstSelector . ',' . $secondSelector;
+        $subject = DeclarationBlock::parse(new ParserState($joinedSelectors . '{}', Settings::create()));
+
+        $resultSelectorStrings = \array_map(
+            static function (Selector $selectorObject): string {
+                return $selectorObject->getSelector();
+            },
+            $subject->getSelectors()
+        );
+        self::assertSame([$firstSelector, $secondSelector], $resultSelectorStrings);
     }
 }